ENH: adds times in integration tests
