+++ backend/prisma/seeds/course.seed.ts
@@ -1,0 +1,7 @@
+import { PrismaClient } from '@prisma/client';
+
+export async function insertCourse(prisma: PrismaClient) {
+  await prisma.course.createMany({
+    data: [],
+  });
+}
