+++ backend/prisma/seeds/seed.ts
@@ -1,11 +1,13 @@
-import { PrismaClient } from '@prisma/client';
 import { insertUser } from '@PrismaORM/seeds/user.seed';
+import { PrismaClient } from '@prisma/client';
+import { insertCourse } from './course.seed';
 
 async function main() {
   const prismaClient = new PrismaClient();
   try {
     await prismaClient.$connect();
     await insertUser(prismaClient);
+    await insertCourse(prismaClient);
   } catch (e) {
     console.log(e);
   }
