+++ frontend/src/components/auth/SignUp.tsx
@@ -9,42 +9,6 @@ const SignUp = () => {
   const navigator = useNavigate();
 
   const handleFormSubmit = async (data: any) => {
-    // toast("Hurray! Account created 🥳🥳", {
-    //   autoClose: 2500,
-    //   closeButton: false,
-    // });
-    console.log("outside fetch");
-
-    // const response = await fetch(
-    //   "https://backend-mu-plum.vercel.app/auth/signup",
-    //   {
-    //     method: "POST",
-    //     headers: {
-    //       "Content-Type": "application/json",
-    //     },
-    //     body: JSON.stringify({ email: data.email, password: data.password }),
-    //   }
-    // );
-
-    // if (response.ok) {
-    //   console.log("success  !!");
-    //   toast("Hurray! Account created 🥳🥳", {
-    //     autoClose: 2500,
-    //     closeButton: false,
-    //   });
-    //   console.log("success!!");
-    //   setTimeout(() => {
-    //     navigator("/auth/signin");
-    //   }, 3000);
-    // } else {
-    //   console.log("fail!!");
-    //   const jsonResponse = await response.json();
-    //   toast.error(jsonResponse.message, {
-    //     autoClose: 2500,
-    //     closeButton: false,
-    //   });
-    // }
-
     const response = await fetch(
       "https://backend-mu-plum.vercel.app/auth/signup",
       {
@@ -54,35 +18,23 @@ const SignUp = () => {
         },
         body: JSON.stringify({ email: data.email, password: data.password }),
       }
-    )
-      .then(async (response) => {
-        if (response.ok) {
-          console.log("success  !!");
-          toast("Hurray! Account created 🥳🥳", {
-            autoClose: 2500,
-            closeButton: false,
-          });
-          console.log("success!!");
-          setTimeout(() => {
-            navigator("/auth/signin");
-          }, 3000);
-        } else {
-          console.log("fail!!");
-          const jsonResponse = await response.json();
-          toast.error(jsonResponse.message, {
-            autoClose: 2500,
-            closeButton: false,
-          });
-        }
-      })
-      .catch((error) => {
-        console.log("fail!!");
-        // const jsonResponse = await response.json();
-        toast.error(error.message, {
-          autoClose: 2500,
-          closeButton: false,
-        });
+    );
+
+    if (response.ok) {
+      toast("Hurray! Account created 🥳🥳", {
+        autoClose: 2500,
+        closeButton: false,
       });
+      setTimeout(() => {
+        navigator("/auth/signin");
+      }, 3000);
+    } else {
+      const jsonResponse = await response.json();
+      toast.error(jsonResponse.message, {
+        autoClose: 2500,
+        closeButton: false,
+      });
+    }
   };
 
   return (
